+++ tests/unit/io/managers/test_context_manager.py
@@ -108,7 +108,7 @@ class TestContextManager(object):
         cm.copy_from(df, 'TABLE NAME', 'replace')
 
         # Then
-        mock.assert_called_once_with('table_name', 'schema', columns, [], True, regenerate=True)
+        mock.assert_called_once_with('table_name', 'schema', columns, [], True)
 
     def test_copy_from_exists_replace_truncate(self, mocker):
         # Given
@@ -124,7 +124,7 @@ class TestContextManager(object):
         cm.copy_from(df, 'TABLE NAME', 'replace')
 
         # Then
-        mock.assert_called_once_with('table_name', 'schema', True, False)
+        mock.assert_called_once_with('table_name', 'schema', True)
 
     def test_internal_copy_from(self, mocker):
         # Given